Update utils.ml
The inferred type of `exit ();` before the change is `unit -> unit`. But (here)[https://github.com/facebook/hhvm/blob/master/hphp/hack/src/utils/sys/sys_utils.ml#L197] the return type of `exit` is `unit -> int`. When compiling with `-strict-sequence`, the compiler will yell:
```
Error: This expression has type int but an expression was expected of type
         unit
```
The fix makes it so that the inferred signature of `exit` becomes `unit -> int`.

The alternative fix here is to ignore the return type `~exit` in `sys_utils.ml` with: `~exit:(fun () -> Unix.umask !old_umask |> ignore)`. Since there's no `.mli` file, I'm not sure what the intended interface is, but my guess is that the intention is to pass the exit code.